refactor(harbor): extract boat placement and harbor utilities to harbor_boats.js and harbor_util.js
